Stack from any order to any set of key names.
Stack on-the-fly by any combination of key names (in increasing order of key values). 
For instance, can make cdp stacks (2D or 3D) without pre-sorting to cdp order. Also allows making shot and receiver stacks (2D or 3D).
Two range-limits-for-stacking can also be specified on two keys. Such as an offset range and/or an azimuth range.